+++ src/tools/globvar.py
